Move unreachability stuff into its own module
