+++ libs/wire-api/src/Wire/API/MLS/Message.hs
@@ -39,17 +39,6 @@ module Wire.API.MLS.Message
     KnownFormatTag (..),
     verifyMessageSignature,
     mkSignedMessage,
-
-    -- * Failed to process
-    UnreachableUsers (UnreachableUsers),
-    unreachableFromList,
-    FailedToProcess (..),
-    failedToSend,
-    failedToSendMaybe,
-    failedToAdd,
-    failedToAddMaybe,
-    failedToRemove,
-    failedToRemoveMaybe,
   )
 where
 
@@ -60,12 +49,8 @@ import Data.Binary
 import Data.Binary.Get
 import Data.Binary.Put
 import qualified Data.ByteArray as BA
-import Data.Id
 import Data.Json.Util
 import Data.Kind
-import Data.List.NonEmpty
-import qualified Data.List.NonEmpty as NE
-import Data.Qualified
 import Data.Schema
 import Data.Singletons.TH
 import qualified Data.Swagger as S
@@ -79,6 +64,7 @@ import Wire.API.MLS.Group
 import Wire.API.MLS.KeyPackage
 import Wire.API.MLS.Proposal
 import Wire.API.MLS.Serialisation
+import Wire.API.Unreachable
 import Wire.Arbitrary (GenericUniform (..))
 
 data WireFormatTag = MLSPlainText | MLSCipherText
@@ -329,90 +315,6 @@ instance SerialiseMLS (MessagePayload 'MLSPlainText) where
   -- We do not need to serialise Commit and Application messages,
   -- so the next case is left as a stub
   serialiseMLS _ = pure ()
-
-newtype UnreachableUsers = UnreachableUsers {unreachableUsers :: NonEmpty (Qualified UserId)}
-  deriving stock (Eq, Show)
-  deriving (A.ToJSON, A.FromJSON, S.ToSchema) via Schema UnreachableUsers
-
-instance Semigroup UnreachableUsers where
-  (UnreachableUsers m) <> (UnreachableUsers n) = UnreachableUsers . NE.nub $ m <> n
-
-instance ToSchema UnreachableUsers where
-  schema =
-    named "UnreachableUsers" $
-      UnreachableUsers
-        <$> unreachableUsers
-          .= nonEmptyArray schema
-
-unreachableFromList :: [Qualified UserId] -> Maybe UnreachableUsers
-unreachableFromList = fmap (UnreachableUsers . NE.nub) . nonEmpty
-
--- | Lists of remote users that could not be processed in a federated action,
--- e.g., a message could not be sent to these remote users.
-data FailedToProcess = FailedToProcess
-  { send :: Maybe UnreachableUsers,
-    add :: Maybe UnreachableUsers,
-    remove :: Maybe UnreachableUsers
-  }
-  deriving (Eq, Show)
-  deriving (A.ToJSON, A.FromJSON, S.ToSchema) via Schema FailedToProcess
-
-instance Semigroup FailedToProcess where
-  ftp1 <> ftp2 =
-    FailedToProcess
-      { send = send ftp1 <> send ftp2,
-        add = add ftp1 <> add ftp2,
-        remove = remove ftp1 <> remove ftp2
-      }
-
-instance Monoid FailedToProcess where
-  mempty = FailedToProcess mempty mempty mempty
-
-failedToProcessObjectSchema :: ObjectSchema SwaggerDoc FailedToProcess
-failedToProcessObjectSchema =
-  FailedToProcess
-    <$> send
-      .= maybe_
-        ( optFieldWithDocModifier
-            "failed_to_send"
-            (description ?~ "List of federated users who could not be reached and did not receive the message")
-            (unnamed schema)
-        )
-    <*> add
-      .= maybe_
-        ( optFieldWithDocModifier
-            "failed_to_add"
-            (description ?~ "List of federated users who could not be reached and be added to a conversation")
-            (unnamed schema)
-        )
-    <*> remove
-      .= maybe_
-        ( optFieldWithDocModifier
-            "failed_to_remove"
-            (description ?~ "List of federated users who could not be reached and be removed from a conversation")
-            (unnamed schema)
-        )
-
-instance ToSchema FailedToProcess where
-  schema = object "FailedToProcess" failedToProcessObjectSchema
-
-failedToSend :: [Qualified UserId] -> FailedToProcess
-failedToSend = failedToSendMaybe . unreachableFromList
-
-failedToSendMaybe :: Maybe UnreachableUsers -> FailedToProcess
-failedToSendMaybe us = mempty {send = us}
-
-failedToAdd :: [Qualified UserId] -> FailedToProcess
-failedToAdd = failedToAddMaybe . unreachableFromList
-
-failedToAddMaybe :: Maybe UnreachableUsers -> FailedToProcess
-failedToAddMaybe us = mempty {add = us}
-
-failedToRemove :: [Qualified UserId] -> FailedToProcess
-failedToRemove = failedToRemoveMaybe . unreachableFromList
-
-failedToRemoveMaybe :: Maybe UnreachableUsers -> FailedToProcess
-failedToRemoveMaybe us = mempty {remove = us}
 
 data MLSMessageSendingStatus = MLSMessageSendingStatus
   { mmssEvents :: [Event],
