+++ services/galley/src/Galley/API/MLS/Propagate.hs
@@ -44,8 +44,8 @@ import Wire.API.Event.Conversation
 import Wire.API.Federation.API
 import Wire.API.Federation.API.Galley
 import Wire.API.Federation.Error
-import Wire.API.MLS.Message
 import Wire.API.Message
+import Wire.API.Unreachable
 
 -- | Propagate a message.
 propagateMessage ::
